Add metric initial system, resources, fix some errors

+++ OliWorkshop.AccountingSys/Models/Resume.cs
@@ -7,7 +7,7 @@ namespace OliWorkshop.AccountingSys.Models
 {
     public class Resume
     {
-        public long Total;
+        public decimal Total;
 
         public decimal AvarageTotal;
